[Tool] make starrocks compilable on gcc12/gcc13/clang18 (#49351)
fix build failure for clang-18+gcc13
fix thrift definition, define the message before refering it
patching rocksdb to fix the forward declaration of incomplete type

+++ be/src/storage/persistent_index.cpp
@@ -3778,7 +3778,7 @@ void PersistentIndex::_get_l2_stat(const std::vector<std::unique_ptr<ImmutableIn
 
                     auto iter = usage_and_size_stat.find(key_size);
                     if (iter == usage_and_size_stat.end()) {
-                        usage_and_size_stat.insert({key_size, {usage, size}});
+                        usage_and_size_stat.insert({static_cast<uint32_t>(key_size), {usage, size}});
                     } else {
                         iter->second.first += usage;
                         iter->second.second += size;
